// src/modules/resellers/interfaces/ResellerRequestInterface.ts
export default interface ResellerRequestInterface {
  A3_COD: string;
  A3_NOME: string;
  A3_END: string;
  A3_BAIRRO: string;
  A3_MUN: string;
  A3_EST: string;
  A3_CEP: string;
  A3_DDDTEL: string;
  A3_TEL: string;
  A3_CGC: string;
  A3_INSCR: string;
  A3_INSCRM: string;
  A3_CARGO: string;
  A3_DDI: string;
  A3_CEL: string;
  A3_XAGN: string;
  A3_XCC: string;
  A3_XLOGIN: string;
  A3_EMAIL: string;
}